Extract createApolloServer to different file
